Revise retry loop
This fixes a bug where we run into wait to exit on SIGTERM during the
retry delay.

It also fixes a bug where a [Clock_ns.every] loop can be duplicated when
a different loop raises and is retried.

It also fixes a bug where sending SIGTERM twice before the tasks finish
causes a crash. It's nice to be able to force exit when a task hangs,
but that is what SIGKILL is for.

Finally, the resulting code is better organized.